feat(test): 重整routine目录测试文件结构
将3个测试文件重组为1个标准命名的测试文件：
- 删除孤立文件: missing_coverage_test.go, empty_stack_coverage_test.go
- 合并到: routine_test.go (协程管理和任务调度)
- 保持98.6%测试覆盖率和37个测试函数
- 涵盖协程池、缓存功能、恐慌恢复等功能
