--- routine/empty_stack_coverage_test.go
@@ -1,260 +1,0 @@
-package routine
-
-import (
-	"runtime/debug"
-	"sync"
-	"testing"
-	"time"
-)
-
-// TestGoWithRecoverEmptyStackScenario specifically targets the empty stack branch
-func TestGoWithRecoverEmptyStackScenario(t *testing.T) {
-	t.Run("empty_stack_coverage", func(t *testing.T) {
-		done := make(chan bool, 1)
-		
-		GoWithRecover(func() error {
-			defer func() { done <- true }()
-			
-			// Try to create a scenario where debug.Stack might return empty
-			// This is extremely rare but we need to test the code path
-			panic("test empty stack scenario")
-		})
-		
-		// Wait for goroutine to complete
-		select {
-		case <-done:
-			// Success - should handle any stack scenario
-		case <-time.After(1 * time.Second):
-			t.Error("Goroutine did not complete within timeout")
-		}
-	})
-
-	t.Run("stack_behavior_test", func(t *testing.T) {
-		// Test to understand stack behavior in different contexts
-		stack := debug.Stack()
-		if len(stack) == 0 {
-			t.Log("debug.Stack() returned empty in normal context")
-		} else {
-			t.Logf("debug.Stack() returned %d bytes in normal context", len(stack))
-		}
-		
-		done := make(chan bool, 1)
-		
-		GoWithRecover(func() error {
-			defer func() { done <- true }()
-			
-			// Test stack in panic context
-			defer func() {
-				if r := recover(); r != nil {
-					stack := debug.Stack()
-					if len(stack) == 0 {
-						t.Log("debug.Stack() returned empty in panic context")
-					} else {
-						t.Logf("debug.Stack() returned %d bytes in panic context", len(stack))
-					}
-					panic(r) // Re-panic to continue the test
-				}
-			}()
-			
-			panic("test stack behavior")
-		})
-		
-		select {
-		case <-done:
-			// Test completed
-		case <-time.After(1 * time.Second):
-			t.Error("Stack behavior test timed out")
-		}
-	})
-	
-	t.Run("attempt_empty_stack_edge_case", func(t *testing.T) {
-		// This test attempts to trigger the len(st) == 0 case
-		// in GoWithRecover function (lines 85-87).
-		// This is extremely difficult to achieve in normal Go code,
-		// as debug.Stack() almost always returns non-empty stack.
-		// The code path exists for defensive programming.
-		
-		done := make(chan bool, 1)
-		
-		GoWithRecover(func() error {
-			defer func() { done <- true }()
-			
-			// Create a panic scenario - the empty stack case is
-			// handled by the runtime internally and is very rare
-			panic("testing edge case for empty stack")
-		})
-		
-		select {
-		case <-done:
-			// Success - the function handles both empty and non-empty stack cases
-		case <-time.After(1 * time.Second):
-			t.Error("Empty stack edge case test timed out")
-		}
-	})
-}
-
-// TestGoWithRecoverComplexPanicScenarios tests various panic scenarios
-func TestGoWithRecoverComplexPanicScenarios(t *testing.T) {
-	scenarios := []struct {
-		name      string
-		panicFunc func()
-	}{
-		{
-			name: "string_panic",
-			panicFunc: func() { panic("string panic") },
-		},
-		{
-			name: "nil_panic", 
-			panicFunc: func() { panic(nil) },
-		},
-		{
-			name: "int_panic",
-			panicFunc: func() { panic(42) },
-		},
-		{
-			name: "struct_panic",
-			panicFunc: func() { panic(struct{ msg string }{msg: "struct panic"}) },
-		},
-	}
-	
-	for _, scenario := range scenarios {
-		t.Run(scenario.name, func(t *testing.T) {
-			done := make(chan bool, 1)
-			
-			GoWithRecover(func() error {
-				defer func() { done <- true }()
-				scenario.panicFunc()
-				return nil
-			})
-			
-			select {
-			case <-done:
-				// Success - panic was recovered
-			case <-time.After(1 * time.Second):
-				t.Errorf("Scenario %s timed out", scenario.name)
-			}
-		})
-	}
-}
-
-// TestGoWithRecoverNestedPanic tests nested panic scenarios
-func TestGoWithRecoverNestedPanic(t *testing.T) {
-	t.Run("nested_panic_scenario", func(t *testing.T) {
-		done := make(chan bool, 1)
-		
-		GoWithRecover(func() error {
-			defer func() { done <- true }()
-			
-			// Create nested panic scenario
-			func() {
-				defer func() {
-					if r := recover(); r != nil {
-						// Re-panic in nested recovery
-						panic("nested panic")
-					}
-				}()
-				panic("original panic")
-			}()
-			
-			return nil
-		})
-		
-		select {
-		case <-done:
-			// Success
-		case <-time.After(1 * time.Second):
-			t.Error("Nested panic scenario timed out")
-		}
-	})
-}
-
-// TestGoWithRecoverConcurrentPanics tests concurrent panic handling
-func TestGoWithRecoverConcurrentPanics(t *testing.T) {
-	t.Run("concurrent_panics", func(t *testing.T) {
-		const numGoroutines = 10
-		var wg sync.WaitGroup
-		
-		for i := 0; i < numGoroutines; i++ {
-			wg.Add(1)
-			
-			GoWithRecover(func() error {
-				defer wg.Done()
-				panic("concurrent panic test")
-			})
-		}
-		
-		done := make(chan struct{})
-		go func() {
-			wg.Wait()
-			close(done)
-		}()
-		
-		select {
-		case <-done:
-			// All panics should be recovered
-			t.Log("All concurrent panics were recovered successfully")
-		case <-time.After(2 * time.Second):
-			t.Error("Concurrent panics test timed out")
-		}
-	})
-}
-
-// TestGoWithRecoverStackManipulation attempts to create edge case scenarios
-func TestGoWithRecoverStackManipulation(t *testing.T) {
-	t.Run("deep_stack_scenario", func(t *testing.T) {
-		done := make(chan bool, 1)
-		
-		GoWithRecover(func() error {
-			defer func() { done <- true }()
-			
-			// Create a deep call stack before panicking
-			var deepFunc func(int)
-			deepFunc = func(depth int) {
-				if depth <= 0 {
-					panic("deep stack panic")
-				}
-				deepFunc(depth - 1)
-			}
-			
-			deepFunc(20) // Create 20-level deep stack
-			return nil
-		})
-		
-		select {
-		case <-done:
-			// Success
-		case <-time.After(1 * time.Second):
-			t.Error("Deep stack scenario timed out")
-		}
-	})
-	
-	t.Run("goroutine_stack_scenario", func(t *testing.T) {
-		done := make(chan bool, 1)
-		
-		GoWithRecover(func() error {
-			defer func() { done <- true }()
-			
-			// Test different stack scenarios within the same goroutine
-			defer func() {
-				if r := recover(); r != nil {
-					// Let GoWithRecover handle this panic
-					panic(r)
-				}
-			}()
-			
-			// Create a scenario that might affect stack traces
-			func() {
-				panic("inner function panic")
-			}()
-			
-			return nil
-		})
-		
-		select {
-		case <-done:
-			// Success
-		case <-time.After(1 * time.Second):
-			t.Error("Goroutine stack scenario timed out")
-		}
-	})
-}
