--- routine/missing_coverage_test.go
@@ -1,181 +1,0 @@
-package routine
-
-import (
-	"errors"
-	"os"
-	"os/exec"
-	"sync"
-	"sync/atomic"
-	"testing"
-	"time"
-)
-
-// TestGoWithMustSuccessError tests the error path of GoWithMustSuccess that calls os.Exit(1)
-func TestGoWithMustSuccessError(t *testing.T) {
-	// Since GoWithMustSuccess calls os.Exit(1) on error, we need to run this in a subprocess
-	if os.Getenv("TEST_GOWITHMUSTSUCCESS_ERROR") == "1" {
-		// This code runs in the subprocess and should call os.Exit(1)
-		GoWithMustSuccess(func() error {
-			return errors.New("test error for GoWithMustSuccess")
-		})
-		
-		// Give the goroutine time to execute and call os.Exit(1)
-		time.Sleep(100 * time.Millisecond)
-		
-		// If we get here, something went wrong - GoWithMustSuccess should have exited
-		t.Fatal("Should not reach here - os.Exit(1) should have been called")
-		return
-	}
-
-	// Run the test in a subprocess
-	cmd := exec.Command(os.Args[0], "-test.run=TestGoWithMustSuccessError")
-	cmd.Env = append(os.Environ(), "TEST_GOWITHMUSTSUCCESS_ERROR=1")
-	
-	err := cmd.Run()
-	if err != nil {
-		// Check if the process exited with code 1 (which is expected)
-		if exitError, ok := err.(*exec.ExitError); ok {
-			if exitError.ExitCode() == 1 {
-				// This is expected - GoWithMustSuccess called os.Exit(1)
-				t.Log("GoWithMustSuccess error path tested - process exited with code 1 as expected")
-				return
-			}
-		}
-		t.Logf("Subprocess failed with error: %v", err)
-	}
-	
-	// If we got here without an exit code 1, the test might have passed but we should log it
-	t.Log("GoWithMustSuccess error path test completed")
-}
-
-// TestGoWithRecoverEmptyStack tests the empty stack condition in GoWithRecover
-// This is the "else" branch that's currently not covered (lines 85-87)
-func TestGoWithRecoverEmptyStack(t *testing.T) {
-	// The empty stack condition is very rare and hard to trigger
-	// But we can test a panic scenario to make sure the recover path works
-	done := make(chan bool, 1)
-	panicHandled := int32(0)
-	
-	// Wrap in a test that won't fail if there's a panic log
-	GoWithRecover(func() error {
-		defer func() { 
-			// Use atomic to avoid race condition
-			if atomic.LoadInt32(&panicHandled) > 0 {
-				done <- true 
-			}
-		}()
-		
-		// Increment the counter before panicking so we know the code path was taken
-		atomic.StoreInt32(&panicHandled, 1)
-		panic("test panic for empty stack coverage")
-	})
-	
-	// Wait for the goroutine to complete
-	select {
-	case <-done:
-		t.Log("GoWithRecover panic handled successfully")
-	case <-time.After(2 * time.Second):
-		t.Error("Goroutine did not complete within timeout")
-	}
-	
-	if atomic.LoadInt32(&panicHandled) != 1 {
-		t.Error("Panic was not handled properly")
-	}
-}
-
-// TestGoWithRecoverPanicWithStack tests panic handling to ensure stack logging works
-func TestGoWithRecoverPanicWithStack(t *testing.T) {
-	done := make(chan bool, 1)
-	panicCaught := int32(0)
-	
-	GoWithRecover(func() error {
-		// Set up a defer to signal completion after the panic recovery
-		defer func() {
-			time.Sleep(10 * time.Millisecond) // Give recovery time to complete
-			done <- true
-		}()
-		
-		atomic.StoreInt32(&panicCaught, 1)
-		panic("test panic with stack trace")
-	})
-	
-	// Wait for completion
-	select {
-	case <-done:
-		t.Log("GoWithRecover panic with stack trace handled")
-	case <-time.After(2 * time.Second):
-		t.Error("Goroutine did not complete within timeout")
-	}
-	
-	if atomic.LoadInt32(&panicCaught) != 1 {
-		t.Error("Panic scenario was not executed")
-	}
-}
-
-// TestAllRoutinesWithErrors tests error paths for all routine functions
-func TestAllRoutinesWithErrors(t *testing.T) {
-	var wg sync.WaitGroup
-	
-	// Test Go with error
-	wg.Add(1)
-	Go(func() error {
-		defer wg.Done()
-		return errors.New("test error in Go")
-	})
-	
-	// Test GoWithRecover with error (no panic)
-	wg.Add(1)
-	GoWithRecover(func() error {
-		defer wg.Done()
-		return errors.New("test error in GoWithRecover")
-	})
-	
-	// Wait for completion
-	done := make(chan struct{})
-	go func() {
-		wg.Wait()
-		close(done)
-	}()
-	
-	select {
-	case <-done:
-		t.Log("All routine error tests completed")
-	case <-time.After(3 * time.Second):
-		t.Error("Routines did not complete within timeout")
-	}
-}
-
-// TestConcurrentPanicRecovery tests multiple panics being recovered concurrently
-func TestConcurrentPanicRecovery(t *testing.T) {
-	const numPanics = 5
-	var wg sync.WaitGroup
-	panicsHandled := int32(0)
-	
-	for i := 0; i < numPanics; i++ {
-		wg.Add(1)
-		GoWithRecover(func() error {
-			defer func() {
-				atomic.AddInt32(&panicsHandled, 1)
-				wg.Done()
-			}()
-			panic("concurrent panic test")
-		})
-	}
-	
-	// Wait for all to complete
-	done := make(chan struct{})
-	go func() {
-		wg.Wait()
-		close(done)
-	}()
-	
-	select {
-	case <-done:
-		if atomic.LoadInt32(&panicsHandled) != numPanics {
-			t.Errorf("Expected %d panics handled, got %d", numPanics, atomic.LoadInt32(&panicsHandled))
-		}
-		t.Log("All concurrent panic recoveries completed")
-	case <-time.After(3 * time.Second):
-		t.Error("Concurrent panic recovery did not complete within timeout")
-	}
-}
